DEEPBLUE-241 - Add ordering to collection subcollection members - 2
